Added reoder function for queue and made swap functions void and pass by reference
I was able to find a way to organize the priority queue so that the best state with lowest h and g values is selected. I also made the swap functions void so that I can make the change in the vector being passed in directly by reference for simplicity purposes.